Merge in FOOD/tele-meet-bot from feature/FOOD-131 to develop

* commit '3d247f5a03b70356490a8a6ab623f39a93181429':
  FOOD-131 all menu
  FOOD-131 add algorithm and test table

+++ src/bot/DBController.py
@@ -85,7 +85,7 @@ class DBController:
             "id": "SERIAL PRIMARY KEY",
             "user_id": "BIGINT",
             "product": "VARCHAR(255)",
-            "type": "BIGINT",
+            "type": "INTEGER",
         }
         self.admin_table_columns = {
             "chat_id": "VARCHAR(255) PRIMARY KEY",
